Updated code with the latest version of the service

+++ Source/UltimateTvGuide.Domain/Entities/TvShow.cs
@@ -7,7 +7,7 @@ namespace UltimateTvGuide.Domain.Entities
         public string Name { get; set; }
         public DateTimeOffset StartTime { get; set; }
         public DateTimeOffset EndTime { get; set; }
-        public float ImdbRating { get; set; }
+        public int ImdbRating { get; set; }
         public string Plot { get; set; }
         public Uri LogoUri { get; set; }
     }
